server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
python-dotenv
requests
psycopg2-binary
gunicorn